% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonpar.R
\name{ase}
\alias{ase}
\title{Compute the Adjacency Spectral embedding}
\usage{
ase(A, d)
}
\arguments{
\item{A}{a symmetric matrix}

\item{d}{the known dimension}
}
\value{
adjacency spectral embedding
}
\description{
Compute the Adjacency Spectral embedding
}
