% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonpar.R
\name{get_dist_matrix}
\alias{get_dist_matrix}
\title{Function to get pairwise distances.}
\usage{
get_dist_matrix(Z1, Z2, sigma = 0.5)
}
\arguments{
\item{Z1}{the n x d dataset}

\item{Z2}{the m x d dataset}

\item{sigma}{the kernel sigma}
}
\value{
the kernel matrix
}
\description{
Function to get pairwise distances.
}
