% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonpar.R
\name{kernel.stat}
\alias{kernel.stat}
\title{Function to generate the U-statistic given the latent positions X and Y
and an optional choice of sigma.
Original code by Youngser Park.}
\usage{
kernel.stat(X, Y, sigma = 0.5, dist = NULL, i1 = c(1:nrow(X)),
  i2 = c((nrow(X) + 1):(nrow(X) * 2)))
}
\description{
Function to generate the U-statistic given the latent positions X and Y
and an optional choice of sigma.
Original code by Youngser Park.
}
