% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonpar.R
\name{nonpar}
\alias{nonpar}
\title{Nonparametric Random Graph Testing}
\usage{
nonpar(A, B, d = 5, nsims = 100, Q_init = NULL)
}
\arguments{
\item{A}{An adjacency matrix}

\item{B}{another adjacency matrix}

\item{d}{the (known) dimension}

\item{nsims}{the number of permutations to run}

\item{Q_init}{initial guess for orthogonal matrix}
}
\description{
Nonparametric Random Graph Testing
}
