% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonpar.R
\name{nonpar.test}
\alias{nonpar.test}
\title{actually run the test}
\usage{
nonpar.test(Xhat, Yhat, nsims = 100, alpha = 0.05)
}
\arguments{
\item{Xhat}{the embedded first graph}

\item{Yhat}{the embedded second graph}

\item{nsims}{the number of permutations to do}

\item{alpha}{the rejection type 1 error}
}
\description{
actually run the test
}
