% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimal_transport.R
\name{optimal_transport}
\alias{optimal_transport}
\title{solve the Optimal transport problem}
\usage{
optimal_transport(X, Y, Q = NULL, lambda = 0.1, eps = 0.01)
}
\arguments{
\item{X}{an n x d matrix of points where each row is a point}

\item{Y}{similar, except possibly a different number of points}

\item{Q}{An orthogonal matrix}

\item{lambda}{the parameter to penalize in sinkhorn divergence}

\item{eps}{the tolerance}
}
\value{
P, the n x m matrix of assignments
}
\description{
Optimally transport according to wasserstein cost
C_{ij} = d(QX_i, Y_j)^2.  this is the usual optimal transport problem
except with an allowance of a parameter Q for an orthogonal matrix.
Give two matrices of dimensions n x d and m x y respectively
}
