% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonpar.R
\name{rect.dist}
\alias{rect.dist}
\title{Function to get the distance.  Original code by Youngser Park.}
\usage{
rect.dist(X, Y)
}
\arguments{
\item{X}{n x d dataset}

\item{Y}{m x dataset}
}
\value{
pairwise kernel n x m matrix
}
\description{
Function to get the distance.  Original code by Youngser Park.
}
