% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonpar.R
\name{run_perm_test}
\alias{run_perm_test}
\title{Runs the permutation test given the value of the U-statistic
the number of repetitions, and the estimated latent positions
Code written in Fall of 2018, so may be buggy}
\usage{
run_perm_test(U, nsims, X, Y, dist.mat = NULL, alpha = 0.05)
}
\arguments{
\item{U}{value of null U statistic}

\item{nsims}{number of simulations}

\item{X}{estimates for latent positions 1}

\item{Y}{estimates for latent positions 2}

\item{dist.mat}{Optional, a distance matrix precalculated already so as not to recalculate.
If null, the function is recursively run to calculate it}

\item{alpha}{the rejection level}
}
\value{
either an estimated p-value under the null or the vector of permuted test statistics
}
\description{
Runs the permutation test given the value of the U-statistic
the number of repetitions, and the estimated latent positions
Code written in Fall of 2018, so may be buggy
}
