% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonpar.R
\name{kernel.stat}
\alias{kernel.stat}
\title{Function to generate the U-statistic given the latent positions X and Y
and an optional choice of sigma.
Original code by Youngser Park.}
\usage{
kernel.stat(X, Y, sigma = 0.5, dist = NULL, i1 = c(1:nrow(X)),
  i2 = c((nrow(X) + 1):(nrow(X) + nrow(Y))))
}
\arguments{
\item{X}{n x d datset}

\item{Y}{m x d dataset}

\item{sigma}{for the kernel}

\item{dist}{optional, the precalculated kernel matrix}

\item{i1}{sets of indices of X to run the permutation test}

\item{i2}{sets of indices of X to run the permutation test}
}
\value{
the kernel statistic evaluated at the specified indices
}
\description{
Function to generate the U-statistic given the latent positions X and Y
and an optional choice of sigma.
Original code by Youngser Park.
}
