% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonpar.R
\name{nonpar}
\alias{nonpar}
\title{Nonparametric Random Graph Testing}
\usage{
nonpar(A, B, d = 5)
}
\arguments{
\item{A}{An adjacency matrix}

\item{B}{another adjacency matrix}

\item{d}{the (known) dimension}
}
\description{
Nonparametric Random Graph Testing
}
