% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonpar.R
\name{nonpar.test}
\alias{nonpar.test}
\title{actually run the test}
\usage{
nonpar.test(Xhat, Yhat, nsims = 100, Q_init = NULL)
}
\arguments{
\item{Xhat}{the embedded first graph}

\item{Yhat}{the embedded second graph}

\item{nsims}{the number of permutations to do}

\item{Q_init}{initial guess for orthogonal matrix}
}
\description{
actually run the test
}
