% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimal_transport.R
\name{procrustes}
\alias{procrustes}
\title{Procrustes}
\usage{
procrustes(X, Y, Pi = NULL)
}
\arguments{
\item{X}{the vectors X}

\item{Y}{the vectors Y}

\item{Pi}{the matrix of transports}
}
\value{
the procrustes solution
}
\description{
Function to find the optimal Q.  Classic Procrustes, but with the allowance
of an assignment matrix.
}
